requirements.txt
torch
torchvision
tokenizers
transformers
tqdm
requests
Pillow
safetensors
huggingface_hub
wandb